requirements.txt
Flask==2.0.2
requests==2.27.1
google-cloud-datastore==2.4.0
gunicorn==20.1.0
